Move BSP jars handling from Scala -> JVM
