+++ src/python/pants/backend/java/bsp/rules.py
@@ -15,6 +15,10 @@ from pants.bsp.util_rules.targets import (
     BSPBuildTargetsMetadataResult,
     BSPCompileRequest,
     BSPCompileResult,
+    BSPDependencyModulesRequest,
+    BSPDependencyModulesResult,
+    BSPDependencySourcesRequest,
+    BSPDependencySourcesResult,
     BSPResourcesRequest,
     BSPResourcesResult,
 )
@@ -24,6 +28,8 @@ from pants.engine.target import FieldSet
 from pants.engine.unions import UnionRule
 from pants.jvm.bsp.compile import _jvm_bsp_compile, jvm_classes_directory
 from pants.jvm.bsp.compile import rules as jvm_compile_rules
+from pants.jvm.bsp.dependencies import _jvm_bsp_dependency_modules, _jvm_bsp_dependency_sources
+from pants.jvm.bsp.dependencies import rules as jvm_dependencies_rules
 from pants.jvm.bsp.resources import _jvm_bsp_resources
 from pants.jvm.bsp.resources import rules as jvm_resources_rules
 from pants.jvm.compile import ClasspathEntryRequestFactory
@@ -150,15 +156,51 @@ async def bsp_java_resources_request(
     return result
 
 
+# -----------------------------------------------------------------------------------------------
+# Dependency Modules / Sources
+# -----------------------------------------------------------------------------------------------
+# Bodies live in `pants.jvm.bsp.dependencies` — the request types below just
+# bind the Java field set, and the @rule bodies are one-line delegations.
+
+
+@dataclass(frozen=True)
+class JavaBSPDependencyModulesRequest(BSPDependencyModulesRequest):
+    field_set_type = JavaMetadataFieldSet
+
+
+@rule
+async def java_bsp_dependency_modules(
+    request: JavaBSPDependencyModulesRequest,
+    build_root: BuildRoot,
+) -> BSPDependencyModulesResult:
+    return await _jvm_bsp_dependency_modules(request, build_root)
+
+
+@dataclass(frozen=True)
+class JavaBSPDependencySourcesRequest(BSPDependencySourcesRequest):
+    field_set_type = JavaMetadataFieldSet
+
+
+@rule
+async def java_bsp_dependency_sources(
+    request: JavaBSPDependencySourcesRequest,
+    build_root: BuildRoot,
+) -> BSPDependencySourcesResult:
+    return await _jvm_bsp_dependency_sources(request, build_root)
+
+
 def rules():
     base_rules = (
         *collect_rules(),
         *jvm_compile_rules(),
+        *jvm_dependencies_rules(),
         *jvm_resources_rules(),
         UnionRule(BSPLanguageSupport, JavaBSPLanguageSupport),
         UnionRule(BSPBuildTargetsMetadataRequest, JavaBSPBuildTargetsMetadataRequest),
         UnionRule(BSPHandlerMapping, JavacOptionsHandlerMapping),
         UnionRule(BSPCompileRequest, JavaBSPCompileRequest),
         UnionRule(BSPResourcesRequest, JavaBSPResourcesRequest),
+        UnionRule(BSPDependencyModulesRequest, JavaBSPDependencyModulesRequest),
+        UnionRule(BSPDependencySourcesRequest, JavaBSPDependencySourcesRequest),
     )
     return (*base_rules, *compute_handler_query_rules(base_rules))
